// packages/figma/src/tokens.generated.ts
export const tomuiFontSizes: ReadonlyArray<FontSizeToken> = [
  { name: "xs", size: 12, lineHeight: 16 },
  { name: "sm", size: 13, lineHeight: 15.2941 },
  { name: "base", size: 14, lineHeight: 21 },
  { name: "lg", size: 16, lineHeight: 24 },
];